feat: add validations in supplies, suppliers and products.

+++ src/Repositories/ProductRepository.cs
@@ -48,6 +48,12 @@ namespace ControlDeVenta_Proy.src.Repositories
             {
                 throw new Exception("Product not found.");
             }
+
+            var productHasSupplies = product.Supplies.Any();
+            if (productHasSupplies)
+            {
+                throw new Exception("You can't delete a product that has supplies.");
+            }
             
             _context.Products.Remove(product);
             await _context.SaveChangesAsync();
